Peter's comments part 4
Refactor error dialog helper functions to EditorUtils.
Add asserts to Document when editor is null
Update docs
